move SimpleContext to protelis-test
