+++ protelis/protelis-test/src/test/resources/genericHood04.pt
@@ -2,7 +2,7 @@
  * EXPECTED_RESULT: 99
  */
 import java.lang.Math.max
-import org.protelis.vm.impl.DummyContext.makeTestField
+import org.protelis.test.infrastructure.SimpleContext.makeTestField
 
 def maxF(a, b) {
     java::lang::Math::max(a, b)
